+++ drupal/sites/all/modules/civicrm/CRM/Core/BAO/EntityTag.php
@@ -2,9 +2,9 @@
 
 /*
  +--------------------------------------------------------------------+
- | CiviCRM version 3.3                                                |
+ | CiviCRM version 4.0                                                |
  +--------------------------------------------------------------------+
- | Copyright CiviCRM LLC (c) 2004-2010                                |
+ | Copyright CiviCRM LLC (c) 2004-2011                                |
  +--------------------------------------------------------------------+
  | This file is a part of CiviCRM.                                    |
  |                                                                    |
@@ -30,7 +30,7 @@
  * This class contains functions for managing Tag(tag) for a contact
  *
  * @package CRM
- * @copyright CiviCRM LLC (c) 2004-2010
+ * @copyright CiviCRM LLC (c) 2004-2011
  * $Id$
  *
  */
@@ -97,6 +97,8 @@ class CRM_Core_BAO_EntityTag extends CRM_Core_DAO_EntityTag
 			
 			//invoke post hook on entityTag
         	require_once 'CRM/Utils/Hook.php';
+            // we are using this format to keep things consistent between the single and bulk operations
+            // so a bit different from other post hooks
         	$object = array( 0 => array( 0 => $params['entity_id'] ), 1 => $params['entity_table'] );
         	CRM_Utils_Hook::post( 'create', 'EntityTag', $params['tag_id'], $object );
         }
@@ -132,13 +134,14 @@ class CRM_Core_BAO_EntityTag extends CRM_Core_DAO_EntityTag
     {
         $entityTag = new CRM_Core_BAO_EntityTag( );
         $entityTag->copyValues( $params );
-        $entityTag->delete( );
-        //return $entityTag;
+        if ( $entityTag->find( true ) ) {
+            $entityTag->delete( );
 		
-		//invoke post hook on entityTag
-        require_once 'CRM/Utils/Hook.php';
-        $object = array( 0 => array( 0 => $params['entity_id'] ), 1 => $params['entity_table'] );
-        CRM_Utils_Hook::post( 'delete', 'EntityTag', $params['tag_id'], $object );
+            //invoke post hook on entityTag
+            require_once 'CRM/Utils/Hook.php';
+            $object = array( 0 => array( 0 => $params['entity_id'] ), 1 => $params['entity_table'] );
+            CRM_Utils_Hook::post( 'delete', 'EntityTag', $params['tag_id'], $object );
+        }
     }
 
     /**
@@ -155,6 +158,8 @@ class CRM_Core_BAO_EntityTag extends CRM_Core_DAO_EntityTag
     static function addEntitiesToTag( &$entityIds, $tagId, $entityTable = 'civicrm_contact' ) {
         $numEntitiesAdded    = 0;
         $numEntitiesNotAdded = 0;
+		$entityIdsAdded      = array();
+		
         foreach ( $entityIds as $entityId ) {
             $tag = new CRM_Core_DAO_EntityTag( );
 
@@ -163,6 +168,7 @@ class CRM_Core_BAO_EntityTag extends CRM_Core_DAO_EntityTag
             $tag->entity_table  = $entityTable;
             if ( ! $tag->find( ) ) {
                 $tag->save( );
+				$entityIdsAdded[] = $entityId;
                 $numEntitiesAdded++;
             } else {
                 $numEntitiesNotAdded++;
@@ -171,7 +177,7 @@ class CRM_Core_BAO_EntityTag extends CRM_Core_DAO_EntityTag
 
         //invoke post hook on entityTag
         require_once 'CRM/Utils/Hook.php';
-        $object = array( $entityIds, $entityTable );
+        $object = array( $entityIdsAdded, $entityTable );
         CRM_Utils_Hook::post('create', 'EntityTag', $tagId, $object );
 
         // reset the group contact cache for all groups
@@ -197,6 +203,8 @@ class CRM_Core_BAO_EntityTag extends CRM_Core_DAO_EntityTag
     {
         $numEntitiesRemoved    = 0;
         $numEntitiesNotRemoved = 0;
+		$entityIdsRemoved      = array();
+		
         foreach ( $entityIds as $entityId ) {
             $tag = new CRM_Core_DAO_EntityTag( );
             
@@ -205,6 +213,7 @@ class CRM_Core_BAO_EntityTag extends CRM_Core_DAO_EntityTag
             $tag->entity_table = $entityTable;
             if (  $tag->find( ) ) {
                 $tag->delete( );
+				$entityIdsRemoved[] = $entityId;
                 $numEntitiesRemoved++;
             } else {
                 $numEntitiesNotRemoved++;
@@ -213,7 +222,7 @@ class CRM_Core_BAO_EntityTag extends CRM_Core_DAO_EntityTag
         
         //invoke post hook on entityTag
         require_once 'CRM/Utils/Hook.php';
-        $object = array( $entityIds, $entityTable );
+        $object = array( $entityIdsRemoved, $entityTable );
         CRM_Utils_Hook::post( 'delete', 'EntityTag', $tagId, $object );
         
         // reset the group contact cache for all groups
